fix(ops): complete horizontal scaling semantics and regression coverage
